+++ blog/gatsby-ssr.js
@@ -1,3 +1,10 @@
-import { wrapRootElement as wrap } from './wrap-root-element'
+import React from 'react';
+import { ThemeProvider } from './src/components/ThemeContext'
 
-export const wrapRootElement = wrap
+export const wrapRootElement = ({ element }) => {
+ return (
+  <ThemeProvider>
+    {element}
+  </ThemeProvider>
+ );
+};
