--- blog/src/components/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { ThemeProvider } from './ThemeContext'
-
-function App({ children }) {
-  return (
-    <ThemeProvider>
-      {children}
-    </ThemeProvider>
-  );
-}
-
-export default App;
